Stage

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Entities/Dokument.cs
@@ -9,7 +9,7 @@ namespace TD.Komercijalno.Contracts.Entities
     public class Dokument
     {
         [Column("VRDOK")]
-        public short VrDok { get; set; }
+        public int VrDok { get; set; }
         [Column("BRDOK")]
         public int BrDok { get; set; }
         [Column("INTBROJ")]
